Allow imperfect rotation matrices to be used
Hi! Thanks for that nice lib! 
I used it in a project where I had to convert rotation matrices output by ARtoolkit into quaternions. My problem is that their matrices are not perfectly clean and have some errors that will for instance turn unit vectors in the correct direction but maybe with a length of 1.000001. 

I needed a way to bypass the tests that check if the matrix provided is really orthogonal. 

I am not sure if you will want to include that workaround in your code as this is pretty niche use, but here it is in case you want it.