Quaternion from Matrix Optional Args
Changes:
- Optionally pass in the rtol and atol arguments from
  pyquaternion to numpys `allclose` and `isclose` functions
  when initializing a Quaternion from a matrix
- Add a unit test with and without optional args
  when initializing a Quaternion from a matrix

Initiliaizing a quaternion from a matrix raises a ValueError
if the provided matrix is not orthogonal. This is because
the default precision used by numpy's allclose function is
1e-08.

Developers should have an option to specify a lower precision
if they choose to and still use all of the utilities in this
library.